requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.2
boto3==1.5.10
botocore==1.8.24
cachetools==2.0.1
cairocffi==0.8.0
celery==4.1.1
chardet==3.0.4
decorator==4.2.1
defusedxml==0.5.0
directory-validators==4.5.1
dj-database-url==0.4.2
Django==1.11.11
-e git+https://github.com/stuaxo/django-admin-ip-restrictor.git@8a24adf18708d8f876bf65317306be5d0113ec2b#egg=django_admin_ip_restrictor
django-amazon-ses==2.0.0
django-bootstrap4==0.0.6
django-clear-cache==0.3
django-countries==5.2
django-countries-plus==1.2.1
django-crispy-forms==1.7.1
django-debug-toolbar==1.9.1
django-extensions==2.0.6
django-ipware==1.1.6
django-markdownx==2.0.22
django-model-utils==3.1.1
django-modelcluster==4.1
django-modeltranslation==0.12.2
django-recaptcha==1.2.0
django-redis==4.8.0
django-reversion==2.0.13
django-storages==1.6.5
django-taggit==0.22.2
django-treebeard==4.2.0
djangorestframework==3.6.4
docutils==0.14
draftjs-exporter==2.0.0
gunicorn==19.7.1
html5lib==0.999999999
httpretty==0.9.4
idna==2.6
ipdb==0.11
ipython==6.3.0
ipython-genutils==0.2.0
jedi==0.11.1
jmespath==0.9.3
Markdown==2.6.11
olefile==0.44
parso==0.1.1
pdfkit==0.6.1
pdfrw==0.4
pexpect==4.4.0
phonenumberslite==7.7.4
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.6.1
pytz==2017.2
raven==6.5.0
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
tinycss2==0.6.1
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wagtail==2.0
wagtail-markdown==0.5a3
-e git+https://github.com/infoportugal/wagtail-modeltranslation.git@a61136e0281279ed213fe90da67ee62e6eb6acfb#egg=wagtail_modeltranslation
wcwidth==0.1.7
-e git+git://github.com/Kozea/WeasyPrint.git@5727e57859b7603d4e06589b22c524456a114b1c#egg=WeasyPrint
-e git+git://github.com/dominno/django-moderation.git@8d98d35d7b39993046997f053d1f72e7a585fb3f#egg=django_moderation
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.1
zenpy==1.1.10
-e git+https://github.com/rollokb/directory-signature-auth@feature/pip10#egg=directory-signature-auth
notifications-python-client==4.8.2
